inline tts debug logger into the client file
